[CInterop] Extract Km->IR transformation to common Native module
`CInteropKlibMetadata2IRTransformer` is going to be used in different
compiler components:
* Kotlin/Native backend (IR linkage)
* KLIB tool (dumping IR and IR signatures)
* KLIB dependencies DAG builder (to be implemented)

So, we need to extract this component to a common Native module.

^KT-86840
