+++ kotlin-native/backend.native/compiler/ir/backend.native/build.gradle.kts
@@ -48,6 +48,7 @@ dependencies {
     implementation(project(":native:binary-options"))
     implementation(project(":compiler:cli:cli-native-klib"))
     implementation(project(":native:native.config"))
+    implementation(project(":native:cinterop.deserialization"))
     implementation(project(":kotlinx-metadata-klib"))
     compileOnly(project(":kotlin-metadata")) // Only to fix IDE reporting unresolved references (KTI-3323).
